no relink, better output

+++ cmd/link.go
@@ -19,24 +19,42 @@ var linkCmd = &cobra.Command{
 	Use:   "link",
 	Short: "Link all files from config.toml",
 	RunE: func(cmd *cobra.Command, args []string) error {
+		internal.LogVerbose(verbose, "Starting link command")
 		internal.LogVerbose(verbose, "Checking for config file")
 		cfg, err := config.Load("config.toml")
 		if err != nil {
 			return err
 		}
+		internal.LogVerbose(verbose, "Config file loaded successfully")
+
+		internal.LogVerbose(verbose, "Evaluating files")
 		for src, dest := range cfg.Files {
 			absSrc, _ := filepath.Abs(src)
 			destPath := expandPath(dest)
 			internal.LogVerbose(verbose, "Checking if %s exists", destPath)
 			if _, err := os.Lstat(destPath); err == nil {
+				internal.LogVerbose(verbose, "Checking to see if it's already symlinked")
+				info, err := os.Lstat(destPath)
+				if err != nil {
+					return err
+				}
+				if info.Mode()&os.ModeSymlink != 0 {
+					internal.LogVerbose(verbose, "%s is already symlinked", destPath)
+					return fmt.Errorf("Symlink %s already exists", destPath)
+				}
 				backupPath := destPath + ".deckbak"
 				if dryRun {
 					fmt.Printf("Would backup %s -> %s\n", destPath, backupPath)
 				} else {
-					os.Rename(destPath, backupPath)
-					internal.LogVerbose(verbose, "Backed up %s -> %s", destPath, backupPath)
-					os.Remove(destPath)
-					internal.LogVerbose(verbose, "Removed %s", destPath)
+					// Check if backup already exists
+					if _, err := os.Stat(backupPath); err == nil {
+						internal.LogVerbose(verbose, "Backing up %s -> %s", destPath, backupPath)
+						os.Rename(destPath, backupPath)
+						internal.LogVerbose(verbose, "Backed up %s -> %s", destPath, backupPath)
+						os.Remove(destPath)
+						internal.LogVerbose(verbose, "Removed %s", destPath)
+					}
+					fmt.Printf("󰑌 Backed up %s to %s\n", destPath, backupPath)
 				}
 			}
 
@@ -44,9 +62,9 @@ var linkCmd = &cobra.Command{
 				fmt.Printf("Would symlink %s -> %s\n", absSrc, destPath)
 			} else {
 				if err := os.Symlink(absSrc, destPath); err != nil {
-					fmt.Printf("❌  %s → %s failed: %v\n", src, destPath, err)
+					fmt.Printf(" %s → %s failed: %v\n", src, destPath, err)
 				} else {
-					fmt.Printf("✅  %s → %s\n", src, destPath)
+					fmt.Printf(" %s → %s\n", src, destPath)
 				}
 			}
 		}
